+++ precompiles/precompiles.go
@@ -8,6 +8,7 @@ import (
 	actprecompile "github.com/warden-protocol/wardenprotocol/precompiles/act"
 	asyncprecompile "github.com/warden-protocol/wardenprotocol/precompiles/async"
 	cmn "github.com/warden-protocol/wardenprotocol/precompiles/common"
+	jsonprecompile "github.com/warden-protocol/wardenprotocol/precompiles/json"
 	slinkyprecompile "github.com/warden-protocol/wardenprotocol/precompiles/slinky"
 	wardenprecompile "github.com/warden-protocol/wardenprotocol/precompiles/warden"
 	actkeeper "github.com/warden-protocol/wardenprotocol/warden/x/act/keeper"
@@ -77,6 +78,13 @@ func NewWardenPrecompiles(
 	precompiles[newAsyncPrecompile.Address()] = newAsyncPrecompile
 
 	eventsRegistry.RegisterEvent("warden.async.v1beta1.EventCreateTask", newAsyncPrecompile.GetCreateTaskEvent)
+
+	newJsonPrecompile, err := jsonprecompile.NewPrecompile(cmn.NewAbiEncoder())
+	if err != nil {
+		return nil, err
+	}
+
+	precompiles[newJsonPrecompile.Address()] = newJsonPrecompile
 
 	return precompiles, nil
 }